+++ src/elements/ModalWindow/ModalWindow.fixture.jsx
@@ -10,6 +10,7 @@ import ModalWindow from './ModalWindow'
 import FormGroup from 'forms/FormGroup'
 import FormButton from 'forms/FormButton'
 import FormSelect from 'forms/FormSelect'
+import FormTextInput from 'forms/FormTextInput'
 // import Menu from 'elements/Menu'
 
 // Demo styles for cosmos engine
@@ -228,10 +229,110 @@ class WithSelect extends React.PureComponent /** @lends @ModalsContainer.prototy
 }
 const withSelect = <WithSelect />
 
+class WithInput extends React.PureComponent /** @lends @ModalsContainer.prototype */ {
+  constructor(props) {
+    super(props)
+    this.state = {
+      open: true,
+      value: '',
+      // loading: true, // Use `useLoader` for ability to controlling Loader
+    }
+    // setTimeout(this.close, 3000)
+    // setInterval(this.toggle, 5000)
+  }
+  open = () => this.setState({ open: true })
+  close = () => this.setState({ open: false })
+  handleOpenState = ({ open }) => this.setState({ open })
+  toggle = () => {
+    this.modalRef && this.modalRef.toggle()
+  }
+  setModalWindowRef = (ref) => {
+    this.modalRef = ref
+  }
+  handleLoaderCancel = () => {
+    console.log('handleLoaderCancel')
+    this.setState(({ loading }) => ({ loading: !loading }))
+  }
+  onChange = ({ value }) => {
+    // console.log('withInput:onChange', value)
+    this.setState({ value })
+  }
+  renderActions() {
+    return (
+      <FormGroup id="Actions" flow fullWidth>
+        <FormGroup id="ActionsMain" flow>
+          <FormButton
+            id="ok"
+            icon="faCheck"
+            text="Ok"
+            theme="primary"
+            inline
+          />
+        </FormGroup>
+        <FormGroup id="ActionsRight" flow align="right">
+          <FormButton
+            id="cancel"
+            icon="faTimes"
+            text="Cancel"
+            inline
+          />
+        </FormGroup>
+      </FormGroup>
+    )
+  }
+  onAction = ({ id }) => {
+    console.log('onAction', id)
+  }
+  render() {
+    const { value, open } = this.state
+    console.log('withInput:render', value)
+    const actions = this.renderActions()
+    // const isElement = React.isValidElement(actions)
+    return (
+      <React.Fragment>
+        <ModalWindow
+          id="simple"
+          theme="info"
+          // icon="faExclamationCircle"
+          open={open}
+          className="ModalWindowClass"
+          contentClassName="XXX"
+          windowClassName="WindowClass"
+          wrapperClassName="WrapperClass"
+          onAction={this.onAction}
+          onClose={this.close}
+          onOpen={this.open}
+          handleOpenState={this.handleOpenState}
+          ref={this.setModalWindowRef}
+          actions={actions}
+          title="WithInput"
+          windowWidth="sm"
+          // leftContent="left" // Left column (with icon visual, eg)
+          showCloseButton
+          autoClose
+          useLoader
+          loading={this.state.loading}
+          // loaderTheme="Dark"
+          handleLoaderCancel={this.handleLoaderCancel}
+        >
+          <FormTextInput value={value} onChange={this.onChange} id="TestInput" fullWidth />
+        </ModalWindow>
+        <FormButton
+          text="Toggle WithInput"
+          onClick={this.toggle}
+          inline
+        />
+      </React.Fragment>
+    )
+  }
+}
+const withInput = <WithInput />
+
 const simple = (
   <ModalWindow
+    error={['error 1', 'error2']}
     id="simple"
-    // open
+    open
   >
     <div className="simpleContainer">
       <div className="simpleStub">simpleStub</div>
@@ -243,4 +344,5 @@ export default {
   simple,
   interactive,
   withSelect,
+  withInput,
 }