+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+- 2020.12.30, 20:11 -- v.0.1.18, PD-3519: Creating & updating publish version tag for each build/publish iteration. Using version for `WebUiCore` dependency entry in `package.json` for target projects.
+- 2020.12.29, 23:53 -- v.0.1.10, Modal errors, FormGroup: noWrap mode, WebUiCoreRoot: autoModalsContainer parameter, ModalPortal/ModalWindow: setPortalNode interface method, ModalWindow Content element styles, ModalWindow error prop (rendering errors section in Content element), ModalsContainer: `usePortal` prop (render in-place or on dom tree topmost level), FormButton: trigger ActionsContext handler even onClick is not defined.
 - 2020.12.28, 04:15 -- ModalPortal: uses `config.modals.containerNode` for register/unregister of modals (eg, catching esc key only for topmost visible modal), added theme styless for wrapper, ModalWindow: fixed animations & bugless vertically centering styles, using `WebUiCoreRoot` as root app wrapper (tried to use modals context for entire components tree).
 - 2020.12.28, 01:13 -- v.0.1.8: Export default library wrapper (`WebUiCoreRoot`).
 - 2020.12.28, 00:25 -- PD-3504: Popup based on `ModalPortal` (`updateGeometry` and other updates `windowDomNode` passed from `ModalPortal`). (With extra debug logging.)
@@ -59,5 +61,5 @@
 - 2020.10.05, 20:38 -- v.0.0.15: Extracted `demoSupport.jsx` routines, fixture selection from url query (parameter `fixture`), added avaialble fixtures list if fixture to display is not specified, added DemoWrapper & demoTitle exportable ficture parameters for demo rendering.
 
 <!--
- @changed 2020.12.28, 04:19
+ @changed 2020.12.30, 20:14
 -->